+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+- feat: add custom "markdownlint" rules: "blanks-around-fenced-divs".
+- fix: set "markdownlint" linting to "on type" by default.
+- chore: remove a `console.log()` statement.
+
 ## 0.14.1 (2025-02-23)
 
 - feat: improve the SVG icon.